requirements.txt
psycopg-binary==3.1.18
fastapi==0.110.0
uvicorn==0.29.0
SQLAlchemy==2.0.29
black==24.3.0
python-multipart==0.0.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
rich==13.7.1
pytest==8.1.1